WORK IN PROGRESS
This script is still in test phase. NOTE: $file_scan_name variable was modified in the subshell. That change might be lost when used in line 73 - to be reviewed.  This script takes DPX sequences and generates fixity details for an individual DPX and a framemd5 manifest for the whole sequence. It further checks conformance to a MediaConch DPX policy before either passing or failing the sequence. Failed sequences are moved to dpx_for_review/ folder. Successful sequences should also be moved to a new location - but for now remain within the same folder making test phase a little easier.